requirements.txt
Flask
python-docx
PyPDF2
python-wordpress-xmlrpc